agents/tools/tools-applications/networking-as-tool/api/fastapi/server/fastapi-prebuilt-agent-poc/requirements.txt
langgraph
langchain
langchain-openai
fastapi[standard]
guardrails-ai
python-dotenv
langchain-community
duckduckgo-search